Add ConfidantHelper file
This is a personal project I started working on to improve my C# skills. It's designed as an aide for the video game Persona 5 Royal, helping players make optimal decisions regarding dialogue events with various characters in the game. It's not the most efficient method and is not finished yet but demonstrates my understanding of C# well.